Add api documentation and testing against api documentation
