+++ .gitignore
@@ -6,3 +6,4 @@ coverage
 .env
 junit.xml
 frontend/build
+openapi3.yaml
