+++ Dockerfile
@@ -1,2 +1,3 @@
 FROM node:12.18
+RUN yarn global add speccy
 WORKDIR /app
