+++ docs/openapi/paths/hello.yaml
@@ -1,0 +1,13 @@
+get:
+  tags:
+    - status
+  operationId: Hello World
+  description: Get the hello world message
+  responses:
+    200:
+      description: Hello world message
+      content:
+        text/html; charset=utf-8:
+          schema:
+            type: string
+            example: Hello from ttadp
