+++ docs/openapi/paths/index.yaml
@@ -1,0 +1,2 @@
+'/hello':
+  $ref: './hello.yaml'
