+++ package.json
@@ -19,11 +19,16 @@
     "lint:ci": "eslint -f eslint-formatter-multiple src",
     "lint:all": "yarn lint:ci && yarn --cwd frontend lint:ci",
     "clean": "rm -rf coverage reports frontend/coverage frontend/reports frontend/build",
+    "api:test": "yarn docs:build && dredd --server 'yarn server'",
+    "docs:build": "npx speccy resolve docs/openapi/index.yaml -o docs/openapi/openapi3.yaml",
+    "docs:lint": "yarn docs:build && npx speccy lint docs/openapi/openapi3.yaml",
+    "docs:serve": "yarn docs:build && npx speccy serve docs/openapi/openapi3.yaml",
     "docker:deps": "docker-compose run --rm backend yarn install && docker-compose run --rm frontend yarn install",
     "docker:start": "docker-compose up -d",
     "docker:stop": "docker-compose down",
     "docker:test": "docker-compose run --rm backend yarn test:ci && docker-compose run --rm frontend yarn test:ci",
     "docker:lint": "docker-compose run --rm backend yarn lint:ci && docker-compose run --rm frontend yarn lint:ci",
+    "docker:api:test": "yarn docs:build && concurrently \"yarn docker:start\" \"sleep 5 && dredd\" && yarn docker:stop",
     "docker:shell:frontend": "docker-compose run --rm frontend /bin/bash",
     "docker:shell:backend": "docker-compose run --rm backend /bin/bash"
   },
@@ -87,20 +92,21 @@
     "jest-cli": "^26.4.2",
     "jest-junit": "^11.1.0",
     "nodemon": "^2.0.4",
+    "speccy": "^0.11.0",
     "supertest": "^4.0.2"
   },
   "dependencies": {
     "axios": "^0.20.0",
     "client-oauth2": "^4.3.3",
     "cors": "^2.8.5",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
     "express-unless": "^0.5.0",
+    "express-winston": "^4.0.5",
     "helmet": "^4.1.0",
     "lodash": "^4.17.20",
     "memorystore": "^1.6.2",
-    "dotenv": "^8.2.0",
-    "express-winston": "^4.0.5",
     "winston": "^3.3.3"
   }
 }
